Architecture - create separation between local and public end points
The GUI now exlusively uses the local end point. This has different authentication clearances which can be utilised in the future. The GUI can open a browser with the correct token info to also connect with the local end point. This browser must be on localhost.

The public end point will be used for all external clients - both localhost and on external machiens with different ip addresses.

There will likely be some rough edges to polish after this commit...