Boot - --with-release-server flag to dev build
This forces the server to be booted in release mode - but also frees the dev build to start the gui with dev tools to inspect and work with a prod server.

Note that tidewave doesn't work in a prod server (by design) but the chromium dev tools and the external gui dev mcp server will work with this.